chore(front): Rename lib/axios client
